SWIP-994 - Refactor and implement Options pattern
