+++ apps/auth-service/Dfe.Sww.Ecf/src/Dfe.Sww.Ecf.AuthorizeAccess/AuthorizeAccessOptions.cs
@@ -3,4 +3,6 @@ namespace Dfe.Sww.Ecf.AuthorizeAccess;
 public class AuthorizeAccessOptions
 {
     public required bool ShowDebugPages { get; set; }
+
+    public string? KeyVaultUri { get; set; }
 }
