+++ apps/auth-service/Dfe.Sww.Ecf/src/Dfe.Sww.Ecf.AuthorizeAccess/Infrastructure/Logging/WebApplicationBuilderExtensions.cs
@@ -1,6 +1,6 @@
 using Dfe.Sww.Ecf.AuthorizeAccess.Infrastructure.Configuration;
+using Microsoft.Extensions.Options;
 using Serilog;
-
 
 namespace Dfe.Sww.Ecf.AuthorizeAccess.Infrastructure.Logging;
 
@@ -10,11 +10,12 @@ public static class WebApplicationBuilderExtensions
     {
         var featureFlags = builder.Services
             .BuildServiceProvider()
-            .GetRequiredService<FeatureFlags>();
+            .GetRequiredService<IOptions<FeatureFlagOptions>>()
+            .Value;
 
         if (featureFlags.EnableSentry)
         {
-            builder.WebHost.UseSentry(dsn: builder.Configuration.GetRequiredValue("Sentry:Dsn"));
+            builder.WebHost.UseSentry(builder.Configuration.GetRequiredValue("Sentry:Dsn"));
         }
 
         builder.Services.AddApplicationInsightsTelemetry();
@@ -22,7 +23,8 @@ public static class WebApplicationBuilderExtensions
         // We want all logging to go through Serilog so that our filters are always applied
         builder.Logging.ClearProviders();
 
-        builder.Host.UseSerilog((ctx, services, config) => config.ConfigureSerilog(ctx.HostingEnvironment, ctx.Configuration, services));
+        builder.Host.UseSerilog((ctx, services, config) =>
+            config.ConfigureSerilog(ctx.HostingEnvironment, ctx.Configuration, services));
 
         return builder;
     }
