+++ apps/auth-service/Dfe.Sww.Ecf/src/Dfe.Sww.Ecf.AuthorizeAccess/Infrastructure/Security/Configuration/OidcOptions.cs
@@ -1,10 +1,14 @@
 namespace Dfe.Sww.Ecf.AuthorizeAccess.Infrastructure.Security.Configuration;
 
-public sealed class OidcConfiguration
+public sealed class OidcOptions
 {
     public const string ConfigurationKey = "Oidc";
 
-    public string? Issuer { get; init; }
+    public required string Issuer { get; init; }
+
+    public string? SigningCertificateName { get; init; }
+
+    public string? EncryptionCertificateName { get; init; }
 
     public OidcApplicationConfiguration[] Applications { get; init; } = [];
 }
